refactor de nombre de carpetas. de NewsActivity ahora es NewsFlow y RSSFeedActivity ahora es RSSFeedFlow
