+++ app/src/main/java/ar/com/thomas/mydailynews/view/NewsFlow/FragmentNewsViewPagerAdapter.java
@@ -1,17 +1,15 @@
-package ar.com.thomas.mydailynews.view.NewsActivity;
+package ar.com.thomas.mydailynews.view.NewsFlow;
 
 import android.content.Context;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
-import android.widget.Toast;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import ar.com.thomas.mydailynews.controller.NewsController;
 import ar.com.thomas.mydailynews.model.News;
-import ar.com.thomas.mydailynews.util.ResultListener;
 
 /**
  * Created by alejandrothomas on 6/25/16.
