+++ local.properties
@@ -7,6 +7,6 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Wed Jun 29 18:35:54 GMT-03:00 2016
-ndk.dir=/Users/mbp/Library/Android/sdk/ndk-bundle
-sdk.dir=/Users/mbp/Library/Android/sdk
+#Thu Jun 30 01:23:09 GMT-03:00 2016
+ndk.dir=/Users/alejandrothomas/Library/Android/sdk/ndk-bundle
+sdk.dir=/Users/alejandrothomas/Library/Android/sdk
